+++ pwncat/commands/upload.py
@@ -13,7 +13,7 @@ from rich.progress import (
 
 import pwncat
 from pwncat.util import console, copyfileobj, human_readable_size, human_readable_delta
-from pwncat.commands import Complete, Parameter, CommandDefinition
+from pwncat.commands import Complete, Parameter, ParseType, CommandDefinition
 
 
 class Command(CommandDefinition):
@@ -23,10 +23,9 @@ class Command(CommandDefinition):
 
     PROG = "upload"
     ARGS = {
-        "source": Parameter(Complete.LOCAL_FILE),
+        "source": Parameter(Complete.LOCAL_FILE, parser=ParseType.LOCAL_FILE),
         "destination": Parameter(
-            Complete.REMOTE_FILE,
-            nargs="?",
+            Complete.REMOTE_FILE, nargs="?", parser=ParseType.REMOTE_FILE
         ),
     }