+++ CHANGELOG.md
@@ -36,6 +36,9 @@ require a major refactor so I'm going to leave it.
 - Added licensing for pwncat (MIT)
 - Added background listener API and commands ([#43](https://github.com/calebstewart/pwncat/issues/43))
 - Added Windows privilege escalation via BadPotato plugin ([#106](https://github.com/calebstewart/pwncat/issues/106))
+- Added command parameter parsers
+- Added homedir (`~`) support on local file completer and parser
+- Added homedir (`~`) support on remote file completer and parser
 ### Removed
 - Removed `setup.py` and `requirements.txt`